requirements.txt
APScheduler==3.9.1
arrow==0.17.0
asgiref==3.5.2
autopep8==1.5.4
class-registry==2.1.2
Django==4.1.2
django-apscheduler==0.6.2
django-crispy-forms==1.10.0
django-filter==2.4.0
django-guardian==2.3.0
djangorestframework==3.12.2
filters==1.3.2
install==1.3.4
Markdown==3.3.3
Pillow==8.4.0
pycodestyle==2.6.0
pycryptodomex==3.9.9
Pygments==2.7.3
PyMySQL==1.0.2
python-dateutil==2.8.1
python-memcached==1.59
pytz==2020.4
pytz-deprecation-shim==0.1.0.post0
regex==2020.11.13
six==1.15.0
sqlparse==0.4.2
toml==0.10.2
tzdata==2022.4
tzlocal==4.2
